# doc/apipost-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "\u62a5\u544a\u751f\u6210\u5e73\u53f0\u63a5\u53e3\u6587\u6863",
    "description": "ReportGenerate \u540e\u7aef\u5df2\u5b9e\u73b0\u63a5\u53e3\uff0c\u5305\u62ec\u7cfb\u7edf\u63a2\u6d3b\u3001\u6587\u4ef6\u7ba1\u7406\u3001\u6a21\u677f\u7ba1\u7406\u548c\u5065\u5eb7\u68c0\u67e5\u3002",
    "version": "0.0.1-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "\u672c\u5730\u5f00\u53d1\u73af\u5883"
    }
  ],
...
  "paths": {
...
    "/api/templates": {
      "post": {
        "tags": [
          "Template"
        ],
        "summary": "\u4e0a\u4f20\u6a21\u677f",
        "description": "\u4e0a\u4f20 Word \u6a21\u677f\uff0c\u4fdd\u5b58\u6a21\u677f\u5143\u6570\u636e\uff0c\u5e76\u89e3\u6790\u6a21\u677f\u53d8\u91cf\u3002\u53c2\u6570\u5fc5\u987b\u653e\u5728 Body \u7684 form-data \u4e2d\u3002",
        "operationId": "createTemplate",
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestIdHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "templateName",
                  "reportType"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Word \u6a21\u677f\u6587\u4ef6\uff1bApipost \u4e2d\u8bf7\u9009\u62e9 File \u7c7b\u578b\u3002"
                  },
                  "templateCode": {
                    "type": "string",
                    "nullable": true,
                    "description": "\u6a21\u677f\u7f16\u7801\uff0c\u4e0d\u586b\u7531\u540e\u7aef\u751f\u6210",
                    "example": "TPL_CRYPTO_001"
                  },
                  "templateName": {
                    "type": "string",
                    "description": "\u6a21\u677f\u540d\u79f0",
                    "example": "\u5bc6\u7801\u6d4b\u8bc4\u6a21\u677f"
                  },
                  "reportType": {
                    "type": "string",
                    "description": "\u62a5\u544a\u7c7b\u578b",
                    "example": "CRYPTO"
                  },
                  "versionNo": {
                    "type": "string",
                    "nullable": true,
                    "description": "\u7248\u672c\u53f7\uff0c\u4e0d\u586b\u9ed8\u8ba4 1.0",
                    "default": "1.0",
                    "example": "1.0"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true,
                    "description": "\u6a21\u677f\u8bf4\u660e",
                    "example": "\u6a21\u677f\u8bf4\u660e"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "\u6210\u529f",
            "headers": {
              "X-Request-Id": {
                "description": "\u8bf7\u6c42\u94fe\u8def ID",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "\u8bf7\u6c42\u53c2\u6570\u9519\u8bef\u6216\u4e1a\u52a1\u6821\u9a8c\u5931\u8d25",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "\u670d\u52a1\u5185\u90e8\u9519\u8bef",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Template"
        ],
        "summary": "\u67e5\u8be2\u6a21\u677f\u5217\u8868",
        "description": "\u5206\u9875\u67e5\u8be2\u6a21\u677f\u5217\u8868\u3002reportType\u3001status\u3001keyword\u3001pageNum\u3001pageSize \u662f Query \u53c2\u6570\uff0c\u4e0d\u653e Body\u3002",
        "operationId": "listTemplates",
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestIdHeader"
          },
          {
            "name": "reportType",
            "in": "query",
            "required": false,
            "description": "\u62a5\u544a\u7c7b\u578b",
            "schema": {
              "type": "string"
            },
            "example": "CRYPTO"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "\u6a21\u677f\u72b6\u6001",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "ENABLED",
                "DISABLED"
              ]
            },
            "example": "DRAFT"
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "description": "\u5173\u952e\u5b57",
            "schema": {
              "type": "string"
            },
            "example": "\u5bc6\u7801"
          },
          {
            "name": "pageNum",
            "in": "query",
            "required": false,
            "description": "\u9875\u7801",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "example": 1
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "description": "\u6bcf\u9875\u6570\u91cf\uff0c\u6700\u5927 100",
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            },
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "\u6210\u529f",
            "headers": {
              "X-Request-Id": {
                "description": "\u8bf7\u6c42\u94fe\u8def ID",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatePageApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "\u8bf7\u6c42\u53c2\u6570\u9519\u8bef\u6216\u4e1a\u52a1\u6821\u9a8c\u5931\u8d25",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "\u670d\u52a1\u5185\u90e8\u9519\u8bef",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/templates/{templateId}": {
      "get": {
        "tags": [
          "Template"
        ],
        "summary": "\u67e5\u8be2\u6a21\u677f\u8be6\u60c5",
        "description": "\u6839\u636e\u6a21\u677f ID \u67e5\u8be2\u6a21\u677f\u5143\u6570\u636e\u548c\u53d8\u91cf\u5217\u8868\u3002templateId \u662f\u8def\u5f84\u53c2\u6570\uff0c\u586b\u5199\u5728 Path \u4e2d\uff0c\u4e0d\u653e Body\u3002",
        "operationId": "getTemplate",
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestIdHeader"
          },
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "description": "\u6a21\u677f ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "\u6210\u529f",
            "headers": {
              "X-Request-Id": {
                "description": "\u8bf7\u6c42\u94fe\u8def ID",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "\u8bf7\u6c42\u53c2\u6570\u9519\u8bef\u6216\u4e1a\u52a1\u6821\u9a8c\u5931\u8d25",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "\u670d\u52a1\u5185\u90e8\u9519\u8bef",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Template"
        ],
        "summary": "\u5220\u9664\u6a21\u677f",
        "description": "\u903b\u8f91\u5220\u9664\u6a21\u677f\u3001\u6a21\u677f\u53d8\u91cf\u548c\u5173\u8054\u6587\u4ef6\u5143\u6570\u636e\uff1b\u4e0d\u7269\u7406\u5220\u9664 MinIO \u6587\u4ef6\u3002",
        "operationId": "deleteTemplate",
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestIdHeader"
          },
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "description": "\u6a21\u677f ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "\u5220\u9664\u6210\u529f",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyApiResponse"
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": null,
                  "requestId": "REQ-APIPOST-DELETE-TEMPLATE"
                }
              }
            }
          },
          "400": {
            "description": "\u8bf7\u6c42\u53c2\u6570\u9519\u8bef\u6216\u4e1a\u52a1\u6821\u9a8c\u5931\u8d25",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "\u670d\u52a1\u5185\u90e8\u9519\u8bef",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/templates/{templateId}/variables": {
      "get": {
        "tags": [
          "Template"
        ],
        "summary": "\u67e5\u8be2\u6a21\u677f\u53d8\u91cf",
        "description": "\u67e5\u8be2\u6a21\u677f\u89e3\u6790\u51fa\u7684\u53d8\u91cf\u6e05\u5355\u3002templateId \u662f\u8def\u5f84\u53c2\u6570\uff0c\u586b\u5199\u5728 Path \u4e2d\uff0c\u4e0d\u653e Body\u3002",
        "operationId": "listTemplateVariables",
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestIdHeader"
          },
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "description": "\u6a21\u677f ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "\u6210\u529f",
            "headers": {
              "X-Request-Id": {
                "description": "\u8bf7\u6c42\u94fe\u8def ID",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateVariableListApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "\u8bf7\u6c42\u53c2\u6570\u9519\u8bef\u6216\u4e1a\u52a1\u6821\u9a8c\u5931\u8d25",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "\u670d\u52a1\u5185\u90e8\u9519\u8bef",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Template"
        ],
        "summary": "\u66ff\u6362\u6a21\u677f\u53d8\u91cf",
        "description": "\u7528\u8bf7\u6c42\u4f53\u4e2d\u7684\u53d8\u91cf\u5217\u8868\u66ff\u6362\u6307\u5b9a\u6a21\u677f\u7684\u5168\u90e8\u53d8\u91cf\u3002templateId \u662f Path \u53c2\u6570\uff1bvariables \u653e\u5728 JSON Body \u4e2d\u3002",
        "operationId": "updateTemplateVariables",
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestIdHeader"
          },
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "description": "\u6a21\u677f ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateVariableUpdateRequest"
              },
              "example": {
                "variables": [
                  {
                    "variableCode": "projectName",
                    "variableName": "\u9879\u76ee\u540d\u79f0",
                    "variableType": "TEXT",
                    "required": true,
                    "defaultValue": "\u67d0\u67d0\u9879\u76ee",
                    "optionsJson": null,
                    "description": "\u9879\u76ee\u540d\u79f0\u5360\u4f4d\u7b26",
                    "sortNo": 1
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "\u6210\u529f",
            "headers": {
              "X-Request-Id": {
                "description": "\u8bf7\u6c42\u94fe\u8def ID",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateVariableListApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "\u8bf7\u6c42\u53c2\u6570\u9519\u8bef\u6216\u4e1a\u52a1\u6821\u9a8c\u5931\u8d25",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "\u670d\u52a1\u5185\u90e8\u9519\u8bef",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/templates/{templateId}/enable": {
      "post": {
        "tags": [
          "Template"
        ],
        "summary": "\u542f\u7528\u6a21\u677f",
        "description": "\u5c06\u6a21\u677f\u72b6\u6001\u8bbe\u7f6e\u4e3a ENABLED\u3002templateId \u662f\u8def\u5f84\u53c2\u6570\uff0c\u586b\u5199\u5728 Path \u4e2d\uff0c\u4e0d\u653e Body\u3002",
        "operationId": "enableTemplate",
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestIdHeader"
          },
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "description": "\u6a21\u677f ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "\u6210\u529f",
            "headers": {
              "X-Request-Id": {
                "description": "\u8bf7\u6c42\u94fe\u8def ID",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoidApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "\u8bf7\u6c42\u53c2\u6570\u9519\u8bef\u6216\u4e1a\u52a1\u6821\u9a8c\u5931\u8d25",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "\u670d\u52a1\u5185\u90e8\u9519\u8bef",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/templates/{templateId}/disable": {
      "post": {
        "tags": [
          "Template"
        ],
        "summary": "\u505c\u7528\u6a21\u677f",
        "description": "\u5c06\u6a21\u677f\u72b6\u6001\u8bbe\u7f6e\u4e3a DISABLED\u3002templateId \u662f\u8def\u5f84\u53c2\u6570\uff0c\u586b\u5199\u5728 Path \u4e2d\uff0c\u4e0d\u653e Body\u3002",
        "operationId": "disableTemplate",
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestIdHeader"
          },
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "description": "\u6a21\u677f ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "\u6210\u529f",
            "headers": {
              "X-Request-Id": {
                "description": "\u8bf7\u6c42\u94fe\u8def ID",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoidApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "\u8bf7\u6c42\u53c2\u6570\u9519\u8bef\u6216\u4e1a\u52a1\u6821\u9a8c\u5931\u8d25",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "\u670d\u52a1\u5185\u90e8\u9519\u8bef",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/templates/{templateId}/download-url": {
      "get": {
        "tags": [
          "Template"
        ],
        "summary": "\u83b7\u53d6\u6a21\u677f\u6587\u4ef6\u4e0b\u8f7d URL",
        "description": "\u6839\u636e\u6a21\u677f\u7ed1\u5b9a\u7684 fileId \u751f\u6210\u6a21\u677f\u6587\u4ef6\u4e34\u65f6\u4e0b\u8f7d URL\uff0c\u4e0d\u76f4\u63a5\u66b4\u9732 MinIO objectName\u3002templateId \u662f Path \u53c2\u6570\uff1bexpireSeconds \u662f Query \u53c2\u6570\u3002",
        "operationId": "getTemplateDownloadUrl",
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestIdHeader"
          },
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "description": "\u6a21\u677f ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "expireSeconds",
            "in": "query",
            "required": false,
            "description": "\u4e34\u65f6 URL \u6709\u6548\u671f\uff0c\u6700\u5927 3600 \u79d2\u3002",
            "schema": {
              "type": "integer",
              "default": 600,
              "minimum": 1,
              "maximum": 3600
            },
            "example": 600
          }
        ],
        "responses": {
          "200": {
            "description": "\u6210\u529f",
            "headers": {
              "X-Request-Id": {
                "description": "\u8bf7\u6c42\u94fe\u8def ID",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateDownloadUrlApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "\u8bf7\u6c42\u53c2\u6570\u9519\u8bef\u6216\u4e1a\u52a1\u6821\u9a8c\u5931\u8d25",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "\u670d\u52a1\u5185\u90e8\u9519\u8bef",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "XRequestIdHeader": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "description": "\u8bf7\u6c42\u94fe\u8def ID\uff1b\u4e0d\u4f20\u65f6\u540e\u7aef\u81ea\u52a8\u751f\u6210\u3002",
        "schema": {
          "type": "string"
        },
        "example": "REQ202606300001"
      }
    },
    "schemas": {
...
      "TemplateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 20
          },
          "templateCode": {
            "type": "string",
            "example": "TPL_CRYPTO_001"
          },
          "templateName": {
            "type": "string",
            "example": "\u5bc6\u7801\u6d4b\u8bc4\u6a21\u677f"
          },
          "reportType": {
            "type": "string",
            "example": "CRYPTO"
          },
          "versionNo": {
            "type": "string",
            "example": "1.0"
          },
          "fileId": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "example": 10001
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "\u6a21\u677f\u8bf4\u660e"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ENABLED",
              "DISABLED"
            ],
            "example": "DRAFT"
          },
          "latest": {
            "type": "boolean",
            "example": true
          },
          "variables": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateVariableResponse"
            }
          }
        }
      },
...
      "TemplateApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/TemplateResponse"
          },
          "requestId": {
            "type": "string",
            "example": "REQ202606300001"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "TemplatePageApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/TemplatePageResult"
          },
          "requestId": {
            "type": "string",
            "example": "REQ202606300001"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "TemplateVariableListApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateVariableResponse"
            }
          },
          "requestId": {
            "type": "string",
            "example": "REQ202606300001"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "VoidApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "nullable": true
          },
          "requestId": {
            "type": "string",
            "example": "REQ202606300001"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "validation failed"
          },
          "data": {
            "nullable": true
          },
          "requestId": {
            "type": "string",
            "example": "REQ202606300001"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "EmptyApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "nullable": true
          },
          "requestId": {
            "type": "string",
            "example": "REQ-APIPOST-DELETE-TEMPLATE"
          }
        }
      },
...
      "TemplateVariableUpdateRequest": {
        "type": "object",
        "required": [
          "variables"
        ],
        "properties": {
          "variables": {
            "type": "array",
            "minItems": 1,
            "description": "\u65b0\u7684\u6a21\u677f\u53d8\u91cf\u5b8c\u6574\u5217\u8868\uff1b\u540e\u7aef\u4f1a\u5148\u903b\u8f91\u5220\u9664\u65e7\u53d8\u91cf\uff0c\u518d\u6279\u91cf\u5199\u5165\u672c\u5217\u8868\u3002",
            "items": {
              "$ref": "#/components/schemas/TemplateVariableUpdateItem"
            }
          }
        }
      },
...
      "TemplateDownloadUrlApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/TemplateDownloadUrlResponse"
          },
          "requestId": {
            "type": "string",
            "example": "REQ202606300001"
          }
        },
        "required": [
          "code",
          "message"
        ]
      }
    }
  }
}